Idiom for left disjoint relations
sig Vehicle says doors: disj  set Door, so each Vehicle has its own particular doors, but we also need each Door to be on only one Vehicle. This seems to be the way to do that. I gloss it thus: Tuples in Vehicle which have _that vehicle's_ doors as (in?) their range are members of the set of tuples from at most one Vehicle to a Door. It also seems to be idiomatic to have such facts free floating, rather than attached to a sig.