Experiment: elide explicit door lock state
It's been suggested in review comments that the model would be simpler if a Vehicle held a set of locked Doors and a set of unlocked Doors, and then we don't need any mechanics.

However, in future we will have the double locked state, and I want to see the state clearly on the visualiser. Let's experiment with that. Current state: not demonstrably inconsistent, but no instances.